+++ domaciDusan/domaciUpisIspis/razbijanjeRecenice.txt
@@ -1,3 +1,7 @@
-bla
-bla
-bla
+Dusan
+i
+Tijana
+uskoro
+6
+godina
+zajedno
